Remove support for tab-delimited annotation files
- use json-schema to validate annotations input file
