+++ .gitignore
@@ -41,3 +41,4 @@ bin
 dag*
 tutorial_data
 reference_inputs
+tmp
