--- tests/data/mock_reference_annotations.full.tsv
@@ -1,6 +1,0 @@
-ensembl_gene_id	hugo_names	chr	strand	gene_start	gene_end	best_ensembl_transcript_id	ensembl_transcript_id	refseq_equivalents	transcript_genomic_start	transcript_genomic_end	cdna_coding_start	cdna_coding_end	genomic_exon_ranges	AA_domain_ranges
-ENSG00000186354	C9orf47	fakereference9	1	1	5278	ENST00000375851	ENST00000375851	NP_001135885.1;NM_001142413.1	1	5278	134	685	1-322;608-833;990-5278	
-ENSG00000186354	C9orf47	fakereference9	1	1	5278	ENST00000375851	ENST00000375850		59	1202	76	783	59-322;608-1202	
-ENSG00000186354	C9orf47	fakereference9	1	1	5278	ENST00000375851	ENST00000334490	NP_001001938.1;NM_001001938.3	66	5278	69	677	66-379;608-833;990-5278	
-ENSG00000213694	S1PR3	fakereference9	1	585	14148	ENST00000375846	ENST00000358157	NP_005217.2;NM_005226.3	585	14148	397	1533	585-833;10192-14148	PR00362:49-62,185-200,298-308;PR00642:63-75,102-112,139-155,329-345;PS50262:56-298;PF10320:55-312;SSF81321:1-340;PR00237:41-65,74-95,118-140,153-174,196-219,241-265,280-306;PR01523:13-25,92-101,112-123,194-204,215-224,272-283,301-311;PF00001:56-298;PR01524:24-40,139-155,223-233,314-323,326-340
-ENSG00000213694	S1PR3	fakereference9	1	585	14148	ENST00000375846	ENST00000375846		5644	14148	4696	5832	5644-14148	PF10320:55-312;PR00362:49-62,185-200,298-308;PS50262:56-298;PR00642:63-75,102-112,139-155,329-345;PR00237:41-65,74-95,118-140,153-174,196-219,241-265,280-306;PR01523:13-25,92-101,112-123,194-204,215-224,272-283,301-311;PR01524:24-40,139-155,223-233,314-323,326-340;PF00001:56-298;SSF81321:1-340
