--- tests/data/mock_reference_annotations.tsv
@@ -1,7 +1,0 @@
-ensembl_gene_id	chr	strand	gene_start	gene_end	ensembl_transcript_id	transcript_genomic_start	transcript_genomic_end
-GENE-A	fake	+	100	200	TRANSCRIPT-A	100	200
-GENE-B	fake	-	250	350	TRANSCRIPT-B	250	350
-GENE-C	fake	+	300	400	TRANSCRIPT-C	300	400
-GENE-D	fake	-	450	550	TRANSCRIPT-D	450	550
-GENE-E	fake	+	500	600	TRANSCRIPT-E	500	600
-GENE-F	fake	+	550	650	TRANSCRIPT-E	550	650
